Update RelWizard.py
Trying to incorporate the code from the excel dashboard workbook generator. I need to figure out how to get all the functions to work properly, i.e. creating the chamber dataframes and having them accesible through the opening screen. Maybe it needs its own separte class? Like "class dashboardGen" or something. I'll have to experiment with it. 

Also I'm stuck on line 256 and 257, I can't figure out how to get this round() function to return what I want. I keep getting an error saying that a string is being passed to it which is incompatible with round() but I can't figure out why it's being passed a string instead of a float. It doesn't produce that error in the separate dashboard generator code.