"Spreadsheet still open" permission loop added
Added a permission loop to try handling the error thrown when the user attempts to update the spreadsheet with the spreadsheet still open in excel.

Initial testing shows the loop to successfully catch the error one time, but additional attempts to update the spreadsheet from the confirmUpdate() prompt throw the error again. I think I need to add something to the except block? I need to find a way to keep looping through the permissionLoop() function while the spreadsheet is still open, close the spreadsheet, then update from the confirmUpdate() prompt